fix: move firebase config to root folder
